+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [3.3.2](https://github.com/typescript-language-server/typescript-language-server/compare/v3.3.1...v3.3.2) (2023-04-17)
+
+
+### Bug Fixes
+
+* formatting fails to provide edit at the end of the document ([#718](https://github.com/typescript-language-server/typescript-language-server/issues/718)) ([01e24a2](https://github.com/typescript-language-server/typescript-language-server/commit/01e24a2cba1b0faf9c65474449265b944e64abd6))
+* verbosity of the "Using Typescript version" message ([#716](https://github.com/typescript-language-server/typescript-language-server/issues/716)) ([448d3f1](https://github.com/typescript-language-server/typescript-language-server/commit/448d3f15d881e2594e1e60f71539b3a5d7e9fce2))
+
+## [3.3.1](https://github.com/typescript-language-server/typescript-language-server/compare/v3.3.0...v3.3.1) (2023-03-27)
+
+
+### Bug Fixes
+
+* don't report InternalError on tsserver error response ([#709](https://github.com/typescript-language-server/typescript-language-server/issues/709)) ([3e63165](https://github.com/typescript-language-server/typescript-language-server/commit/3e6316546eb5c8b6fd2fb8c26c88b7b6a6331472))
+
 ## [3.3.0](https://github.com/typescript-language-server/typescript-language-server/compare/v3.2.0...v3.3.0) (2023-02-20)